implement trait for rpc+cache
