+++ lib/client/src/account_fetcher_utils.rs
@@ -1,3 +1,4 @@
+use anchor_lang::AccountDeserialize;
 use anyhow::Context;
 use solana_sdk::account::ReadableAccount;
 use solana_sdk::pubkey::Pubkey;
@@ -13,4 +14,16 @@ pub async fn account_fetcher_fetch_mango_account(
     let data: &[u8] = &account.data();
     MangoAccountValue::from_bytes(&data[8..])
         .with_context(|| format!("deserializing mango account {}", address))
+}
+
+
+// Can't be in the trait, since then it would no longer be object-safe...
+pub async fn account_fetcher_fetch_anchor_account<T: AccountDeserialize>(
+    fetcher: &dyn AccountFetcher,
+    address: &Pubkey,
+) -> anyhow::Result<T> {
+    let account = fetcher.fetch_raw_account(address).await?;
+    let mut data: &[u8] = &account.data();
+    T::try_deserialize(&mut data)
+        .with_context(|| format!("deserializing anchor account {}", address))
 }
