+++ lib/client/src/lib.rs
@@ -5,7 +5,7 @@ pub use util::*;
 
 mod account_fetcher_trait;
 mod account_fetchers;
-mod mango_account_fetchers;
+mod account_fetcher_utils;
 pub mod account_update_stream;
 pub mod chain_data;
 mod chain_data_fetcher;
